backend/requirements.txt

flask==2.3.3
flask-cors==4.0.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
alembic==1.12.0
SQLAlchemy==2.0.20
pydantic==2.4.2
